Fix miscellaneous issues
Using ((retry_count++)) in sg_persist_with_retry() would cause the test
to fail. Since retry_count is initially 0, the value of that expression
is 0 the first time it is evaluated. Autoincrement uses arithmetic
evaluation, so a value of zero is treated as a return status of 1. use
pre-increment instead to make sure the expression evaluates to a
non-zero number.

Use the device name and not the path for device2, to match device1.